+++ aldrin/CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Add `low_level::Event::new()`.
+
+### Changed
+
+- Move `Event` to the `low_level` module.
+
 ## [0.4.0] - 2024-03-21
 
 This version still connects to brokers using protocol version 14.